Refactor toom-cook 4
