+++ include/beman/big_int/detail/mul_impl.hpp
@@ -297,447 +297,11 @@ constexpr std::size_t toom_cook_4_storage_size(const std::size_t s) noexcept { r
 // `toom_cook_4_cutoff` for this call only. Production callers should omit it so
 // the default cutoff applies. Recursive sub-product calls always use the default.
 // ---------------------------------------------------------------------------
-constexpr void multiply_toom_cook_4(const std::span<uint_multiprecision_t>       result,
-                                    const std::span<const uint_multiprecision_t> a_untrimmed,
-                                    const std::span<const uint_multiprecision_t> b_untrimmed,
-                                    scratch_allocator_base&                scratch,
-                                    const std::size_t                            cutoff_override = 0) noexcept {
-    BEMAN_BIG_INT_DEBUG_ASSERT(!a_untrimmed.empty());
-    BEMAN_BIG_INT_DEBUG_ASSERT(!b_untrimmed.empty());
-    BEMAN_BIG_INT_DEBUG_ASSERT(result.size() >= trimmed_size_span(a_untrimmed) + trimmed_size_span(b_untrimmed));
-    BEMAN_BIG_INT_DEBUG_ASSERT(result.data() != a_untrimmed.data());
-    BEMAN_BIG_INT_DEBUG_ASSERT(result.data() != b_untrimmed.data());
-
-    const auto a = a_untrimmed.first(trimmed_size_span(a_untrimmed));
-    const auto b = b_untrimmed.first(trimmed_size_span(b_untrimmed));
-
-    // Partition at k = ceil(max(an, bn) / 4).
-    const std::size_t min_size         = std::min(a.size(), b.size());
-    const std::size_t k                = (std::max(a.size(), b.size()) + 3) / 4;
-    const std::size_t effective_cutoff = cutoff_override == 0 ? toom_cook_4_cutoff : cutoff_override;
-
-    // Fall through to Toom-Cook 3 (and on through Karatsuba and schoolbook) when the
-    // smaller operand is below the performance cutoff or below the algorithm's 3*k
-    // invariant (Toom-Cook 4 needs both a3 and b3 non-empty).
-    if (min_size < effective_cutoff || min_size <= 3 * k) {
-        multiply_toom_cook_3(result, a, b, scratch);
-        return;
-    }
-
-    // Split each operand into four pieces.
-    const auto a0 = a.first(std::min(k, a.size()));
-    const auto a1 = a.size() > k ? a.subspan(k, std::min(k, a.size() - k)) : std::span<const uint_multiprecision_t>{};
-    const auto a2 =
-        a.size() > 2 * k ? a.subspan(2 * k, std::min(k, a.size() - 2 * k)) : std::span<const uint_multiprecision_t>{};
-    const auto a3 = a.size() > 3 * k ? a.subspan(3 * k) : std::span<const uint_multiprecision_t>{};
-
-    const auto b0 = b.first(std::min(k, b.size()));
-    const auto b1 = b.size() > k ? b.subspan(k, std::min(k, b.size() - k)) : std::span<const uint_multiprecision_t>{};
-    const auto b2 =
-        b.size() > 2 * k ? b.subspan(2 * k, std::min(k, b.size() - 2 * k)) : std::span<const uint_multiprecision_t>{};
-    const auto b3 = b.size() > 3 * k ? b.subspan(3 * k) : std::span<const uint_multiprecision_t>{};
-
-    // Carve scratch:
-    //   tmpa, tmpb: k+2 limbs each (evaluation buffers, hold p(x) and q(x) at each point)
-    //   v1, vm1, v2, vm2, vh: 2k+2 limbs each (five interpolation buffers)
-    //   tmp_double: 2k+2 limbs (used for in-place bit shifts to replace doubling with mul)
-    // c0 = a0*b0 lives in result[0..2k); c6 = a3*b3 lives in result[6k..).
-    const std::size_t tmp_cap       = k + 2;
-    const std::size_t prod_cap      = 2 * k + 2;
-    const std::size_t total_scratch = 2 * tmp_cap + 6 * prod_cap;
-
-    auto block      = scratch.allocate(total_scratch);
-    auto tmpa       = block.first(tmp_cap);
-    auto tmpb       = block.subspan(tmp_cap, tmp_cap);
-    auto v1         = block.subspan(2 * tmp_cap, prod_cap);
-    auto vm1        = block.subspan(2 * tmp_cap + prod_cap, prod_cap);
-    auto v2         = block.subspan(2 * tmp_cap + 2 * prod_cap, prod_cap);
-    auto vm2        = block.subspan(2 * tmp_cap + 3 * prod_cap, prod_cap);
-    auto vh         = block.subspan(2 * tmp_cap + 4 * prod_cap, prod_cap);
-    auto tmp_double = block.subspan(2 * tmp_cap + 5 * prod_cap, prod_cap);
-
-    // ---- c0 = a0*b0, written into result[0..2k) (caller pre-zeroed). ----
-    multiply_toom_cook_4(result.first(a0.size() + b0.size()), a0, b0, scratch);
-
-    // ---- c6 = a3*b3, written into result[6k..). ----
-    // The min > 3*k gate above guarantees both a3 and b3 are non-empty.
-    multiply_toom_cook_4(result.subspan(6 * k, a3.size() + b3.size()), a3, b3, scratch);
-
-    // ---- Evaluate at x = 1: tmpa = a0 + a1 + a2 + a3; tmpb similarly. ----
-    std::ranges::copy(a0, tmpa.begin());
-    std::size_t tmpa_size = a0.size();
-    tmpa_size             = add_into_tmp(tmpa, tmpa_size, a1);
-    tmpa_size             = add_into_tmp(tmpa, tmpa_size, a2);
-    tmpa_size             = add_into_tmp(tmpa, tmpa_size, a3);
-
-    std::ranges::copy(b0, tmpb.begin());
-    std::size_t tmpb_size = b0.size();
-    tmpb_size             = add_into_tmp(tmpb, tmpb_size, b1);
-    tmpb_size             = add_into_tmp(tmpb, tmpb_size, b2);
-    tmpb_size             = add_into_tmp(tmpb, tmpb_size, b3);
-
-    // v1 = tmpa * tmpb
-    std::ranges::fill(v1, uint_multiprecision_t{0});
-    multiply_toom_cook_4(v1,
-                         std::span<const uint_multiprecision_t>{tmpa.data(), tmpa_size},
-                         std::span<const uint_multiprecision_t>{tmpb.data(), tmpb_size},
-                         scratch);
-
-    // ---- Evaluate at x = -1: tmpa = (a0 + a2) - (a1 + a3) signed; tmpb similarly. ----
-    std::ranges::copy(a0, tmpa.begin());
-    tmpa_size = a0.size();
-    tmpa_size = add_into_tmp(tmpa, tmpa_size, a2);
-    // tmpb temporarily used to hold a1 + a3 before subtraction.
-    std::ranges::copy(a1, tmpb.begin());
-    std::size_t aux_size = a1.size();
-    aux_size             = add_into_tmp(tmpb, aux_size, a3);
-    const auto sub_a_m1 =
-        subtract_unsigned_spans_signed(tmpa,
-                                       std::span<const uint_multiprecision_t>{tmpa.data(), tmpa_size},
-                                       std::span<const uint_multiprecision_t>{tmpb.data(), aux_size});
-    tmpa_size            = sub_a_m1.size;
-    const bool sign_a_m1 = sub_a_m1.negative;
-
-    std::ranges::copy(b0, tmpb.begin());
-    tmpb_size = b0.size();
-    tmpb_size = add_into_tmp(tmpb, tmpb_size, b2);
-    // tmpa now holds (a0 + a2) - (a1 + a3); reuse tail of vm1 as a scratch slot for (b1 + b3).
-    std::ranges::copy(b1, vm1.begin());
-    aux_size = b1.size();
-    aux_size = add_into_tmp(vm1, aux_size, b3);
-    const auto sub_b_m1 =
-        subtract_unsigned_spans_signed(tmpb,
-                                       std::span<const uint_multiprecision_t>{tmpb.data(), tmpb_size},
-                                       std::span<const uint_multiprecision_t>{vm1.data(), aux_size});
-    tmpb_size            = sub_b_m1.size;
-    const bool sign_b_m1 = sub_b_m1.negative;
-    const bool sign_vm1  = sign_a_m1 ^ sign_b_m1;
-
-    std::ranges::fill(vm1, uint_multiprecision_t{0});
-    if (tmpa_size != 0 && tmpb_size != 0) {
-        multiply_toom_cook_4(vm1,
-                             std::span<const uint_multiprecision_t>{tmpa.data(), tmpa_size},
-                             std::span<const uint_multiprecision_t>{tmpb.data(), tmpb_size},
-                             scratch);
-    }
-
-    // ---- Evaluate at x = 2: tmpa = ((a3*2 + a2)*2 + a1)*2 + a0 (Horner); tmpb similarly. ----
-    std::ranges::copy(a3, tmpa.begin());
-    tmpa_size = a3.size();
-    tmpa_size = shift_left_one(tmpa, tmpa_size);
-    tmpa_size = add_into_tmp(tmpa, tmpa_size, a2);
-    tmpa_size = shift_left_one(tmpa, tmpa_size);
-    tmpa_size = add_into_tmp(tmpa, tmpa_size, a1);
-    tmpa_size = shift_left_one(tmpa, tmpa_size);
-    tmpa_size = add_into_tmp(tmpa, tmpa_size, a0);
-
-    std::ranges::copy(b3, tmpb.begin());
-    tmpb_size = b3.size();
-    tmpb_size = shift_left_one(tmpb, tmpb_size);
-    tmpb_size = add_into_tmp(tmpb, tmpb_size, b2);
-    tmpb_size = shift_left_one(tmpb, tmpb_size);
-    tmpb_size = add_into_tmp(tmpb, tmpb_size, b1);
-    tmpb_size = shift_left_one(tmpb, tmpb_size);
-    tmpb_size = add_into_tmp(tmpb, tmpb_size, b0);
-
-    std::ranges::fill(v2, uint_multiprecision_t{0});
-    multiply_toom_cook_4(v2,
-                         std::span<const uint_multiprecision_t>{tmpa.data(), tmpa_size},
-                         std::span<const uint_multiprecision_t>{tmpb.data(), tmpb_size},
-                         scratch);
-
-    // ---- Evaluate at x = -2: tmpa = (a0 + 4*a2) - (2*a1 + 8*a3) signed; tmpb similarly.
-    // Build positive = a0 + 4*a2 in tmpa, negative = 2*a1 + 8*a3 in tmpb, then signed-sub.
-    std::ranges::copy(a2, tmpa.begin());
-    tmpa_size = a2.size();
-    tmpa_size = shift_left_n(tmpa, tmpa_size, 2u);
-    tmpa_size = add_into_tmp(tmpa, tmpa_size, a0);
-
-    std::ranges::copy(a3, tmpb.begin());
-    aux_size = a3.size();
-    aux_size = shift_left_n(tmpb, aux_size, 3u);
-    aux_size = add_into_tmp(tmpb, aux_size, a1);
-    // tmpb holds 8*a3 + a1; we still need 2*a1, which is (8*a3 + a1) + a1 = 8*a3 + 2*a1.
-    aux_size = add_into_tmp(tmpb, aux_size, a1);
-
-    const auto sub_a_m2 =
-        subtract_unsigned_spans_signed(tmpa,
-                                       std::span<const uint_multiprecision_t>{tmpa.data(), tmpa_size},
-                                       std::span<const uint_multiprecision_t>{tmpb.data(), aux_size});
-    tmpa_size            = sub_a_m2.size;
-    const bool sign_a_m2 = sub_a_m2.negative;
-
-    std::ranges::copy(b2, tmpb.begin());
-    tmpb_size = b2.size();
-    tmpb_size = shift_left_n(tmpb, tmpb_size, 2u);
-    tmpb_size = add_into_tmp(tmpb, tmpb_size, b0);
-
-    // Use vm2 as a scratch slot for 8*b3 + 2*b1.
-    std::ranges::copy(b3, vm2.begin());
-    aux_size = b3.size();
-    aux_size = shift_left_n(vm2, aux_size, 3u);
-    aux_size = add_into_tmp(vm2, aux_size, b1);
-    aux_size = add_into_tmp(vm2, aux_size, b1);
-
-    const auto sub_b_m2 =
-        subtract_unsigned_spans_signed(tmpb,
-                                       std::span<const uint_multiprecision_t>{tmpb.data(), tmpb_size},
-                                       std::span<const uint_multiprecision_t>{vm2.data(), aux_size});
-    tmpb_size            = sub_b_m2.size;
-    const bool sign_b_m2 = sub_b_m2.negative;
-    const bool sign_vm2  = sign_a_m2 ^ sign_b_m2;
-
-    std::ranges::fill(vm2, uint_multiprecision_t{0});
-    if (tmpa_size != 0 && tmpb_size != 0) {
-        multiply_toom_cook_4(vm2,
-                             std::span<const uint_multiprecision_t>{tmpa.data(), tmpa_size},
-                             std::span<const uint_multiprecision_t>{tmpb.data(), tmpb_size},
-                             scratch);
-    }
-
-    // ---- Evaluate at x = 1/2 (scaled by 8): tmpa = ((a0*2 + a1)*2 + a2)*2 + a3 (Horner);
-    // this equals 8*a(1/2), so the resulting product is 64*c(1/2). ----
-    std::ranges::copy(a0, tmpa.begin());
-    tmpa_size = a0.size();
-    tmpa_size = shift_left_one(tmpa, tmpa_size);
-    tmpa_size = add_into_tmp(tmpa, tmpa_size, a1);
-    tmpa_size = shift_left_one(tmpa, tmpa_size);
-    tmpa_size = add_into_tmp(tmpa, tmpa_size, a2);
-    tmpa_size = shift_left_one(tmpa, tmpa_size);
-    tmpa_size = add_into_tmp(tmpa, tmpa_size, a3);
-
-    std::ranges::copy(b0, tmpb.begin());
-    tmpb_size = b0.size();
-    tmpb_size = shift_left_one(tmpb, tmpb_size);
-    tmpb_size = add_into_tmp(tmpb, tmpb_size, b1);
-    tmpb_size = shift_left_one(tmpb, tmpb_size);
-    tmpb_size = add_into_tmp(tmpb, tmpb_size, b2);
-    tmpb_size = shift_left_one(tmpb, tmpb_size);
-    tmpb_size = add_into_tmp(tmpb, tmpb_size, b3);
-
-    std::ranges::fill(vh, uint_multiprecision_t{0});
-    multiply_toom_cook_4(vh,
-                         std::span<const uint_multiprecision_t>{tmpa.data(), tmpa_size},
-                         std::span<const uint_multiprecision_t>{tmpb.data(), tmpb_size},
-                         scratch);
-
-    // ---- Bodrato-style interpolation (in-place; full 2k+2 buffers throughout) ----
-    // After this sequence the c-coefficients live as:
-    //   c0 in result[0..2k), c1 in vh, c2 in v1, c3 in vm1, c4 in v2, c5 in vm2, c6 in result[6k..).
-    //
-    // v0_view and vinf_view bound the c0 and c6 regions in result. Clip vinf_view because
-    // the result span may be larger than 8k when the caller's prod_cap exceeds the actual
-    // product (same trick as Toom-3 at line 642).
-    const auto v0_view   = std::span<const uint_multiprecision_t>{result.data(), 2 * k};
-    const auto vinf_size = std::min(result.size() - 6 * k, 2 * k);
-    const auto vinf_view = std::span<const uint_multiprecision_t>{result.data() + 6 * k, vinf_size};
-    const auto v1_view   = std::span<const uint_multiprecision_t>{v1};
-    const auto vm1_view  = std::span<const uint_multiprecision_t>{vm1};
-    const auto v2_view   = std::span<const uint_multiprecision_t>{v2};
-    const auto vm2_view  = std::span<const uint_multiprecision_t>{vm2};
-    const auto vh_view   = std::span<const uint_multiprecision_t>{vh};
-    const auto td_view   = std::span<const uint_multiprecision_t>{tmp_double};
-
-    // Phase 1: Symmetrize {v1, vm1} into {E1, D1} and {v2, vm2} into {E2, D2}.
-    //   E1 = (v1 + vm1)/2 = c0 + c2 + c4 + c6
-    //   D1 = (v1 - vm1)/2 = c1 + c3 + c5
-    //   E2 = (v2 + vm2)/2 = c0 + 4c2 + 16c4 + 64c6
-    //   D2 = (v2 - vm2)/4 = c1 + 4c3 + 16c5
-
-    // Step 1a: v1 <- (v1 + vm1) algebraic (sign-aware on sign_vm1).
-    if (sign_vm1) {
-        subtract_unsigned_spans(v1, v1_view, vm1_view);
-    } else {
-        const bool carry_out = add_unsigned_spans(v1, v1_view, vm1_view);
-        BEMAN_BIG_INT_DEBUG_ASSERT(!carry_out);
-    }
-    // Step 1b: v1 /= 2.
-    {
-        const auto rem = shift_right_one(v1);
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-    // After Step 1: v1 = E1 = c0 + c2 + c4 + c6.
-
-    // Step 2: vm1 <- v1 - vm1 algebraic = (v1_orig - vm1_orig)/2 = D1.
-    if (sign_vm1) {
-        const bool carry_out = add_unsigned_spans(vm1, v1_view, vm1_view);
-        BEMAN_BIG_INT_DEBUG_ASSERT(!carry_out);
-    } else {
-        subtract_unsigned_spans(vm1, v1_view, vm1_view);
-    }
-    // After Step 2: vm1 = D1 = c1 + c3 + c5. sign_vm1 is no longer needed.
-
-    // Step 3a: v2 <- (v2 + vm2) algebraic.
-    if (sign_vm2) {
-        subtract_unsigned_spans(v2, v2_view, vm2_view);
-    } else {
-        const bool carry_out = add_unsigned_spans(v2, v2_view, vm2_view);
-        BEMAN_BIG_INT_DEBUG_ASSERT(!carry_out);
-    }
-    // Step 3b: v2 /= 2.
-    {
-        const auto rem = shift_right_one(v2);
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-    // After Step 3: v2 = E2 = c0 + 4c2 + 16c4 + 64c6.
-
-    // Step 4a: vm2 <- v2 - vm2 algebraic = (v2_orig - vm2_orig)/2 = 2c1 + 8c3 + 32c5.
-    if (sign_vm2) {
-        const bool carry_out = add_unsigned_spans(vm2, v2_view, vm2_view);
-        BEMAN_BIG_INT_DEBUG_ASSERT(!carry_out);
-    } else {
-        subtract_unsigned_spans(vm2, v2_view, vm2_view);
-    }
-    // Step 4b: vm2 /= 2.
-    {
-        const auto rem = shift_right_one(vm2);
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-    // After Step 4: vm2 = D2 = c1 + 4c3 + 16c5. sign_vm2 is no longer needed.
-
-    // Phase 2: Solve even system for c2 (into v1) and c4 (into v2).
-
-    // Step 5: v2 -= v1.  v2 = (c0+4c2+16c4+64c6) - (c0+c2+c4+c6) = 3*(c2 + 5c4 + 21c6).
-    subtract_unsigned_spans(v2, v2_view, v1_view);
-    // Step 6: v2 /= 3.  v2 = c2 + 5c4 + 21c6.
-    {
-        const auto rem = divide_unsigned_short(v2, v2_view, uint_multiprecision_t{3});
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-
-    // Step 7: v1 -= c0.  v1 = c2 + c4 + c6.
-    subtract_unsigned_spans(v1, v1_view, v0_view);
-    // Step 8: v1 -= c6.  v1 = c2 + c4.
-    subtract_unsigned_spans(v1, v1_view, vinf_view);
-
-    // Step 9: v2 -= v1.  v2 = 4c4 + 21c6.
-    subtract_unsigned_spans(v2, v2_view, v1_view);
-
-    // Step 10: subtract 21*c6 from v2 using tmp_double for the doublings.
-    //   v2 -= 1*c6, then v2 -= 4*c6, then v2 -= 16*c6.  Total subtracted = 21*c6.
-    subtract_unsigned_spans(v2, v2_view, vinf_view);
-    std::ranges::fill(tmp_double, uint_multiprecision_t{0});
-    std::ranges::copy(vinf_view, tmp_double.begin());
-    std::size_t td_size = vinf_view.size();
-    td_size             = shift_left_n(tmp_double, td_size, 2u);
-    subtract_unsigned_spans(v2, v2_view, std::span<const uint_multiprecision_t>{tmp_double.data(), td_size});
-    td_size = shift_left_n(tmp_double, td_size, 2u);
-    subtract_unsigned_spans(v2, v2_view, std::span<const uint_multiprecision_t>{tmp_double.data(), td_size});
-    // After Step 10: v2 = 4c4.
-
-    // Step 11: v2 /= 4 (two halvings).  v2 = c4.
-    {
-        const auto rem = shift_right_n(v2, 2u);
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-
-    // Step 12: v1 -= v2.  v1 = c2.
-    subtract_unsigned_spans(v1, v1_view, v2_view);
-
-    // Phase 3: Reduce vh to T_odd = (vh - 64c0 - 16c2 - 4c4 - c6) / 2 = 16c1 + 4c3 + c5.
-
-    // Step 13: vh -= 64*c0 (subtract 64*c0 via doubled-tmp_double).
-    std::ranges::fill(tmp_double, uint_multiprecision_t{0});
-    std::ranges::copy(v0_view, tmp_double.begin());
-    td_size = trimmed_size_span(v0_view);
-    td_size = shift_left_n(tmp_double, td_size, 6u);
-    subtract_unsigned_spans(vh, vh_view, std::span<const uint_multiprecision_t>{tmp_double.data(), td_size});
-
-    // Step 14: vh -= 16*c2 (c2 lives in v1 now).
-    std::ranges::fill(tmp_double, uint_multiprecision_t{0});
-    std::ranges::copy(v1_view, tmp_double.begin());
-    td_size = trimmed_size_span(v1_view);
-    td_size = shift_left_n(tmp_double, td_size, 4u);
-    subtract_unsigned_spans(vh, vh_view, std::span<const uint_multiprecision_t>{tmp_double.data(), td_size});
-
-    // Step 15: vh -= 4*c4 (c4 lives in v2 now).
-    std::ranges::fill(tmp_double, uint_multiprecision_t{0});
-    std::ranges::copy(v2_view, tmp_double.begin());
-    td_size = trimmed_size_span(v2_view);
-    td_size = shift_left_n(tmp_double, td_size, 2u);
-    subtract_unsigned_spans(vh, vh_view, std::span<const uint_multiprecision_t>{tmp_double.data(), td_size});
-
-    // Step 16: vh -= c6.
-    subtract_unsigned_spans(vh, vh_view, vinf_view);
-
-    // Step 17: vh /= 2.  vh = 16c1 + 4c3 + c5 = T_odd.
-    {
-        const auto rem = shift_right_one(vh);
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-
-    // Phase 4: Solve odd system using vm1 = D1, vm2 = D2, vh = T_odd.
-    //   vm2 -= D1:  3c3 + 15c5
-    //   vh  -= D1: 15c1 +  3c3
-    //   then divide both by 3.
-
-    // Step 18: vm2 -= vm1.  vm2 = 3*(c3 + 5c5).
-    subtract_unsigned_spans(vm2, vm2_view, vm1_view);
-    // Step 19: vh -= vm1.  vh = 3*(5c1 + c3).
-    subtract_unsigned_spans(vh, vh_view, vm1_view);
-    // Step 20: vm2 /= 3.  vm2 = alpha = c3 + 5c5.
-    {
-        const auto rem = divide_unsigned_short(vm2, vm2_view, uint_multiprecision_t{3});
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-    // Step 21: vh /= 3.  vh = beta = 5c1 + c3.
-    {
-        const auto rem = divide_unsigned_short(vh, vh_view, uint_multiprecision_t{3});
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-
-    // Now vm1 = D1, vm2 = alpha, vh = beta. Recover c3 = (5*D1 - alpha - beta) / 3 into vm1.
-    // Step 22: tmp_double = D1; vm1 *= 4; vm1 += tmp_double  (-> 5*D1).
-    std::ranges::fill(tmp_double, uint_multiprecision_t{0});
-    std::ranges::copy(vm1_view, tmp_double.begin());
-    const auto sz_2D1 = shift_left_one(vm1, vm1.size()); // 2*D1
-    BEMAN_BIG_INT_DEBUG_ASSERT(sz_2D1 == vm1.size());
-    const auto sz_4D1 = shift_left_one(vm1, vm1.size()); // 4*D1
-    BEMAN_BIG_INT_DEBUG_ASSERT(sz_4D1 == vm1.size());
-    {
-        const bool carry = add_unsigned_spans(vm1, vm1_view, td_view); // 5*D1
-        BEMAN_BIG_INT_DEBUG_ASSERT(!carry);
-    }
-    // Step 23: vm1 -= alpha; vm1 -= beta.  vm1 = 3*c3.
-    subtract_unsigned_spans(vm1, vm1_view, vm2_view);
-    subtract_unsigned_spans(vm1, vm1_view, vh_view);
-    // Step 24: vm1 /= 3.  vm1 = c3.
-    {
-        const auto rem = divide_unsigned_short(vm1, vm1_view, uint_multiprecision_t{3});
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-
-    // Step 25: vh -= c3 -> 5*c1; vh /= 5.  vh = c1.
-    subtract_unsigned_spans(vh, vh_view, vm1_view);
-    {
-        const auto rem = divide_unsigned_short(vh, vh_view, uint_multiprecision_t{5});
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-
-    // Step 26: vm2 -= c3 -> 5*c5; vm2 /= 5.  vm2 = c5.
-    subtract_unsigned_spans(vm2, vm2_view, vm1_view);
-    {
-        const auto rem = divide_unsigned_short(vm2, vm2_view, uint_multiprecision_t{5});
-        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
-    }
-
-    // Phase 5: Recompose.
-    //   result += vh * B^k        (c1)
-    //   result += v1 * B^(2k)     (c2)
-    //   result += vm1 * B^(3k)    (c3)
-    //   result += v2 * B^(4k)     (c4)
-    //   result += vm2 * B^(5k)    (c5)
-    // c0 and c6 are already in result.
-    // Carries propagate up to result.size().
-    add_shifted(result, k, vh_view);
-    add_shifted(result, 2 * k, v1_view);
-    add_shifted(result, 3 * k, vm1_view);
-    add_shifted(result, 4 * k, v2_view);
-    add_shifted(result, 5 * k, vm2_view);
-
-    // Release scratch back to the bump pool for sibling reuse.
-    scratch.deallocate(total_scratch);
-}
+void multiply_toom_cook_4(const std::span<uint_multiprecision_t>       result,
+                          const std::span<const uint_multiprecision_t> a_untrimmed,
+                          const std::span<const uint_multiprecision_t> b_untrimmed,
+                          scratch_allocator_base&                      scratch,
+                          const std::size_t                            cutoff_override = 0) noexcept;
 
 // See tests/beman/big_int/perf crossover_speedup.png
 inline constexpr std::size_t toom_cook_6_5_cutoff = 3000;
